requirements.txt
streamlit
google-generativeai>=0.8.6
python-dotenv
pandas
plotly
requests